+++ Change_log.md
@@ -1,9 +1,9 @@
 # Here are the futur **implementations/changes/patches** for the **2.1.1** update:
 ### Additions:
-- Play_mode button in help. `Not DONE`
+- Play_mode button in help. `DONE`
 ***
 ### Changes:
 - None. *Cool!*
 ***
 ### Patches:
-- Fix unintended indentation in About. `Not DONE`
+- Fix unintended indentation in About. `DONE`